continue fixing ticket:1202
- now rnav is also a <ul></ul> list.
- current is a class and not id
- core module href use &amp; instad of &
- rnav will not be generated with empty <li>

basically the core module should be clean enough. i also tried not to break any new things in non patched modules. the only problem, is that id="current" does not make the current item bold anymore. my next goal will be fixing the needed modules, so this is not a big problem on the long term.

module maintainers: please make my life easier and work for me
:)