Fix unit tests for dahdi channels
The unit tests for dahdi channels were broken. This commit makes them
work again. Also, there was an issue with HTML being slipped into output
in an unexpected way, so refactored the component to allow errors to be
passed into the main class. Note, this is probably still not the correct
way to do things as there may still be an issue with concerns, but it is
closer to correct without being too risky of a change.