Remove unused function
While working on a project, my IDE discovered that there was a bug
in sql statement executed in the core_routing_renamebyid function
that would prevent it from executing successfully. Instead of fixing
the bug, I discovered that this function isn't even used anywhere.
So removing the function entirely as the function doesn't work, and
it's been this way for a while without complaint.

No issue was created for this and a release is not necessary.