Refactored code for changing a device tech to sharable method
Previously, the code for changing a device was a combination of things
living in the UI as well as in the form handling component. Because of
this, the logic for changing a device tech was not sharable and we could
not use it to create a command line option for converting devices from
chan_sip to chan_pjsip. This commit pulls out all of the logic and puts
it into a shareable method on the Core Object, which lets us now create
a command line tool.