FREEI-1284: added feature to send out an email when an Outbound Route is dialed
This will add a Notifications tab when adding/modifying and Outbound Route. It will have fields for Email from, to, subject, and body.  The subject and body are template style, and call varibles are provided so people can customize it.  I just stole the emailing method from another module, let me know if I need to do that differently.  I don't know if I need to do something special with smtp settings, or if those will correctly get applied automatically.
I also plan to move the Paging "Notifications" option to this tab, and rename it to "Paging Notifications" to give it some distinction.

For the asterisk diaplan part.
I have doubts that I did this correctly.  I added it to the Dial function in dialout-trunk because that was the only way I could get it to only send an email to the connecting trunk, if two trunks before it failed.  It felt wrong to be passing so many arguments to my macro, but that's the only way I knew how to get all the info I needed in my macro.  Maybe there's a cleaner way.  Any suggestions on any thing is appreciated.

NOTE: bugfix/FREEI-1284(no "_FPBX*") is an old branch that I messed up, and should be disregarded.