FREEPBX-21437: optimized outbound route notification email script usage
This should reduce the amount of calls to the email notification agi script.  The original behavior was an AGI was called every time an outbound call was placed over a trunk.  The AGI script would do some db calls to get email content info, and determine if an email notification should be sent. The new behavior here is that the route's dialplan func will be flagged in a way where the agi will not run at all if an emailto address is not set. The db lookup for emailto happens during a Apply Config/Reload instead of during a call.  Some other minor changes include some code cleanups, and placing a space before AM/PM in the email content.