FREEPBX-21437: optimized outbound route notification email script usage, FPBX13 changes
These are the same changes as the FPBX14,15 PRs regarding the recent changes.  Since 13's PR for the High CPU fixes was not merged yet, I also added the fixes for the outbound routes issue from FREEPBX-21450 to this branch.

This should reduce the amount of calls to the email notification agi script. The original behavior was an AGI was called every time an outbound call was placed over a trunk. The AGI script would do some db calls to get email content info, and determine if an email notification should be sent.

The new behavior here is that the route's dialplan func will be flagged in a way where the agi will not run at all if an emailto address is not set. The db lookup for emailto happens during a Apply Config/Reload instead of during a call.

Some other minor changes include setting some asterisk variables in a better way so they don't have to be passed in to the emailing dialplan func, as well as some code cleanups, and placing a space before AM/PM in the email content.