Update Core.class.php removing hard coding of pjsip dial
PJSIP's dial value was hardcoded to "PJSIP/3" for all extensions when $data('dial') was not set. An existing module we have that created extensions had a conflict after an update due to this. Other properties, such as 'mailbox' on the next line have a precedent for setting the default value with the new extension.